Switching from UInt64 to UInt128 in global interface/mortar id computation (II).

+++ src/solvers/dgsem_t8code/containers_parallel.jl
@@ -21,8 +21,8 @@ function reinitialize_containers!(mesh::ParallelT8codeMesh, equations, dg::DGSEM
                      mpi_interfaces = mpi_interfaces,
 
                      # Temporary arrays for updating `mpi_cache`.
-                     global_mortar_ids = fill(UInt64(0), nmpimortars(mpi_mortars)),
-                     global_interface_ids = fill(UInt64(0), nmpiinterfaces(mpi_interfaces)),
+                     global_mortar_ids = fill(UInt128(0), nmpimortars(mpi_mortars)),
+                     global_interface_ids = fill(UInt128(0), nmpiinterfaces(mpi_interfaces)),
                      neighbor_ranks_mortar = Vector{Vector{Int}}(undef,
                                                                  nmpimortars(mpi_mortars)),
                      neighbor_ranks_interface = fill(-1, nmpiinterfaces(mpi_interfaces)))
